+++ packages/sdks-tests/playwright.config.ts
@@ -23,7 +23,7 @@ const WEB_SERVERS: Record<Sdk, PackageName[]> = {
 
     'e2e-qwik-city',
   ],
-  react: ['e2e-nextjs-react', 'e2e-react'],
+  react: ['e2e-nextjs-react', 'e2e-react', 'e2e-nextjs-app-dir-react'],
   vue: ['e2e-vue2', 'e2e-vue3'],
   svelte: ['e2e-svelte', 'e2e-sveltekit'],
   rsc: [],
@@ -35,7 +35,7 @@ targetContext.name = sdk;
 const isReactNative = sdk === 'reactNative';
 
 const things = WEB_SERVERS[sdk].map((packageName, i) => {
-  const port = isReactNative ? 19006 : 1234 + i;
+  const port = isReactNative ? 19006 : 1111 + i;
   const portFlag = isReactNative ? '' : `--port=${port}`;
 
   return {