+++ packages/sdks-tests/src/tests/sdk.ts
@@ -20,6 +20,7 @@ export type PackageName =
   | 'e2e-qwik'
   | 'e2e-qwik-city'
   | 'e2e-nextjs-react'
+  | 'e2e-nextjs-app-dir-react'
   | 'e2e-react'
   | 'e2e-vue2'
   | 'e2e-vue3'