+++ scripts/merge-dependabot-prs.mjs
@@ -13,19 +13,18 @@ import Octokit from '@octokit/rest';
 import { question } from 'zx';
 import { echo } from 'zx/experimental';
 
-// load the GITHUB_TOKEN from .env file
+// load the GITHUB_TOKEN_X from .env file
 require('dotenv').config();
 
 // authenticate Octokit with a personal access token
 const octokit = new Octokit({
-  auth: process.env.GITHUB_TOKEN,
+  auth: process.env.GITHUB_TOKEN_X,
 });
 
 console.log('Welcome to the BuilderIO/builder dependabot PR merger!');
-console.log(`Using GITHUB_TOKEN: ${process.env.GITHUB_TOKEN}`);
 
-if (!process.env.GITHUB_TOKEN) {
-  throw new Error(`GITHUB_TOKEN not found in .env file`);
+if (!process.env.GITHUB_TOKEN_X) {
+  throw new Error(`GITHUB_TOKEN_X not found in .env file`);
 }
 
 let query = '';
@@ -36,13 +35,30 @@ async function getQuery() {
 
 const getPRs = async ({ isApproved }) => {
   const getQuery = isApproved =>
-    `repo:BuilderIO/builder is:open state:open author:app/dependabot type:pr ${query} in:title review:${
-      isApproved ? 'approved' : 'required'
-    }`;
+    `repo:BuilderIO/builder is:open state:open author:app/dependabot type:pr ${query} in:title `;
   const { data: pullRequests } = await octokit.search.issuesAndPullRequests({
     q: getQuery(isApproved),
+    per_page: 100,
   });
+
   return pullRequests;
+};
+
+/**
+ *
+ * @param {Octokit.SearchIssuesAndPullRequestsResponseItemsItem} pr
+ */
+const autoMerge = async pr => {
+  try {
+    await $`gh pr review ${pr.number} --approve`;
+    // enable auto-merge
+    await $`gh pr merge ${pr.number} --auto --squash`;
+    console.log(`Enabled auto-merge for ${pr.html_url}.`);
+  } catch (e) {
+    console.log(`Error auto-merging ${pr.html_url}.`);
+    // print details
+    console.log(e);
+  }
 };
 
 async function approve() {
@@ -56,37 +72,45 @@ async function approve() {
     console.log(`${pr.html_url} || ${pr.title}`);
   });
 
-  if (unapprovedPullRequests.total_count > 0) {
-    // Confirm with the user
-    let confirm = await question(
-      `Are you sure you want to approve these ${unapprovedPullRequests.total_count} pull requests? (yes/no): `
-    );
-
-    if (confirm !== 'yes') {
-      throw new Error(`Script aborted.`);
-    }
-
-    // Approve all PRs
-    for (const pr of unapprovedPullRequests.items) {
-      // approve the PR and print appropriate message
-      console.log(`Approving ${pr.html_url}...`);
-      const resp = await octokit.pulls.createReview({
-        owner: 'BuilderIO',
-        repo: 'builder',
-        pull_number: pr.number,
-        event: 'APPROVE',
-      });
-      if (resp.status >= 300) {
-        console.log(`Error approving ${pr.html_url}.`);
-        // print details
-        console.log(resp);
-      } else {
-        console.log(`Approved ${pr.html_url}.`);
-      }
-    }
-  } else {
+  if (unapprovedPullRequests.total_count === 0) {
     echo`No unapproved PRs found, skipping approval step`;
+    return;
   }
+
+  // Confirm with the user
+  let confirm = await question(
+    `Are you sure you want to approve these ${unapprovedPullRequests.total_count} pull requests? (yes/no): `
+  );
+
+  if (confirm !== 'yes') {
+    throw new Error(`Script aborted.`);
+  }
+
+  /**
+   *
+   * @param {Octokit.SearchIssuesAndPullRequestsResponseItemsItem} pr
+   */
+  const processPr = async pr => {
+    autoMerge(pr);
+    // approve the PR and print appropriate message
+    console.log(`Approving ${pr.html_url}...`);
+
+    const resp = await octokit.pulls.createReview({
+      owner: 'BuilderIO',
+      repo: 'builder',
+      pull_number: pr.number,
+      event: 'APPROVE',
+    });
+    if (resp.status >= 300) {
+      console.log(`Error approving ${pr.html_url}.`);
+      // print details
+      console.log(resp);
+    } else {
+      console.log(`Approved ${pr.html_url}.`);
+    }
+  };
+
+  await Promise.all(unapprovedPullRequests.items.map(processPr));
 }
 
 async function merge() {
@@ -120,8 +144,7 @@ async function merge() {
    */
   const dependabotRebasePRs = [];
 
-  // Merge all PRs
-  for (const pr of approvedPullRequests.items) {
+  const mergePr = async pr => {
     // merge the PR and print appropriate message
     console.log(`Merging ${pr.html_url}`);
     try {
@@ -148,7 +171,10 @@ async function merge() {
         dependabotRebasePRs.push(pr);
       }
     }
-  }
+  };
+
+  // Merge all PRs
+  await Promise.all(approvedPullRequests.items.map(mergePr));
 
   if (dependabotRebasePRs.length > 0) {
     console.log(`The following PRs need to be rebased by Dependabot: `);
@@ -203,7 +229,29 @@ async function rebaseDependabot(prUrl) {
 async function main() {
   await getQuery();
   await approve();
-  await merge();
+  // await merge();
+}
+
+async function workflowStuff() {
+  let x = null;
+  octokit.actions
+    .listRepoWorkflowRuns({ owner: 'BuilderIO', repo: 'builder', branch: 'main' })
+    .then(k => (x = k));
+
+  const cancelWorkflowRun2 = async workflow => {
+    try {
+      await octokit.request('POST /repos/BuilderIO/builder/actions/runs/{run_id}/cancel', {
+        run_id: workflow.id,
+      });
+      console.log('cancelled workflow:', workflow.id);
+    } catch (error) {
+      console.error('could not cancel workflow:', workflow.id);
+    }
+  };
+
+  octokit
+    .request('GET /repos/BuilderIO/builder/actions/runs?status=queued&branch=main&per_page=10')
+    .then(x => x.data.workflow_runs.map(cancelWorkflowRun2));
 }
 
 main();